application/src/main/java/br/com/payment/application/exception/NoResourceFoundException.java
package br.com.payment.application.exception;

import org.springframework.http.HttpStatus;

public class NoResourceFoundException extends ApiErrorException {
    private static final HttpStatus HTTP_STATUS = HttpStatus.NOT_FOUND;
    private String message;
    public NoResourceFoundException(String message) {
        this.message = message;
    }

    public NoResourceFoundException() {
        super();
        this.message = HTTP_STATUS.getReasonPhrase();
    }

    public int getCode() {
        return HTTP_STATUS.value();
    }

    @Override
    public String getMessage() {
        return message;
    }
}
